# Faster read; faster lookup; more tests