Previously adding (reading a file consisting of) n non-overlapping
packages took O(n^2) time.
Looking up a single package in a structure with n non-overlapping
packages took O(n) time.

Here this is changed so the timings are more like O(n) and O(1),
respectively.

(all of these should be taken with a grain of salt as not only the
number of packages influence the time, but also the length of the paths
which I completely ignore here).

Run like this:
```
$ for i in 1 10 100 1000 5000 10000 50000; do dart test/bench.dart $i;
done
```

(With no warmup, only run once etc --- there's lots of variations
between runs, but for this purpose it doesn't really matter.)

Before:
Read file with 1 packages in 18 ms, looked up all packages in 0 ms
Read file with 10 packages in 22 ms, looked up all packages in 0 ms
Read file with 100 packages in 28 ms, looked up all packages in 1 ms
Read file with 1000 packages in 78 ms, looked up all packages in 14 ms
Read file with 5000 packages in 442 ms, looked up all packages in 384 ms
Read file with 10000 packages in 2254 ms, looked up all packages in 1826 ms
Read file with 50000 packages in 67572 ms, looked up all packages in 84050 ms

After:
Read file with 1 packages in 24 ms, looked up all packages in 0 ms
Read file with 10 packages in 23 ms, looked up all packages in 0 ms
Read file with 100 packages in 25 ms, looked up all packages in 1 ms
Read file with 1000 packages in 60 ms, looked up all packages in 6 ms
Read file with 5000 packages in 127 ms, looked up all packages in 10 ms
Read file with 10000 packages in 187 ms, looked up all packages in 13 ms
Read file with 50000 packages in 525 ms, looked up all packages in 61 ms

Furthermore:
* Previously at least one test didn't work because of a json syntax
  error. This has been fixed.

Also, the code for making a package inside the package root of another package was added, but disabled, for internal reasons. Tests are added an modeled after the old (prev this PR) responses.