@natebosch 2022-01-07
It looks like this was published without being tested in google3 or the SDK. We strongly prefer that packages are synced internally for validation before they are published.  https://github.com/dart-lang/sdk/wiki/External-Package-Maintenance#publishing-a-package