@jensjoha 2022-01-10
What do you mean with "published"? The PR was merged. To my knowledge that's it --- and for instance it doesn't figure on pub.dev (https://pub.dev/packages/package_config says latest version published 3 months ago).