@natebosch 2022-01-25
> and for instance it doesn't figure on pub.dev (https://pub.dev/packages/package_config says latest version published 3 months ago).

Ah, I wasn't looking close enough at the publish date and I was only checking the version listed in the pubspec and changelog here. It looks like it was just that we didn't pre-emptively bump the version to a `-dev` like we typically do.

https://github.com/dart-lang/sdk/wiki/External-Package-Maintenance#making-a-change